requirements.txt
Django==4.2.27
django-allauth==0.58.2
django-bootstrap5==25.1
django-crispy-forms==2.4
crispy-bootstrap5==0.7
django-widget-tweaks==1.5.0
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==23.0.0
Pillow==10.3.0

# Database
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# Project specific
django-tables2==2.7.5
django-environ==0.12.0
django-celery-beat==2.8.1
django-celery-results==2.6.0
django-compressor==4.5.1
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-timezone-field==7.1
python-dateutil==2.9.0
pytz==2024.1
celery==5.4.0
redis==5.0.3
social-auth-app-django==5.4.3
python-dotenv==1.0.1
reportlab==4.1.0
